fix(index.html): ensure the og:image is the icon, not the hero image
previously, was linked to the web_page_ of the imgur hosting of the hero image, and not the address of the image itself. now, it is just the relative path of the branding svg. this allows other people to see the logo of intellex academy whenever intellex.co is shared on social media i think!