refactor: language switching and handling cta buttons and their ensuing behaviours (incl the modal/dialog) are now js modules. fixes #23
language switching should be reusable across all pages. we now detect, in main.js, whether we are on the homepage or not. if not, then we do not attach event listeners to elements (the cta buttons) that don't actually exists. thereby avoiding console errors that appear on those policy pages, which fixes #23